fix(attention): accept uint32 seq_lens in plan()
seq_lens is documented as uint32, but torch has no CPU comparison or
max() for uint32. The built-in max() already failed for two or more
requests, and Tensor.max() fails for one. Convert seq_lens to int32 once
when it enters plan(); the buffers it feeds are int32 already, and the
conversion is a no-op for int32 input.

The plan-length tests now run with int32 and uint32 seq_lens.

+++ flashinfer/decode.py
@@ -1306,7 +1306,7 @@ class BatchDecodeWithPagedKVCacheWrapper:
         if seq_lens is None:
             kv_lens_arr_host = get_seq_lens(indptr_host, last_page_len_host, page_size)
         else:
-            kv_lens_arr_host = seq_lens.cpu()
+            kv_lens_arr_host = seq_lens.cpu().to(torch.int32)
         if q_len_per_req > 1:
             min_kv_len = int(kv_lens_arr_host.min())
             if min_kv_len < q_len_per_req:
@@ -1714,7 +1714,7 @@ class BatchDecodeWithPagedKVCacheWrapper:
         if seq_lens is None:
             kv_lens_arr_host = get_seq_lens(indptr_host, last_page_len_host, page_size)
         else:
-            kv_lens_arr_host = seq_lens.cpu()
+            kv_lens_arr_host = seq_lens.cpu().to(torch.int32)
         if q_len_per_req > 1 and is_causal:
             min_kv_len = int(kv_lens_arr_host.min())
             if min_kv_len < q_len_per_req:
